refactored import-service
